requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-environ==0.8.1
idna==3.3
phonenumbers==8.12.40
pycodestyle==2.8.0
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
